read file descriptor
