+++ lib/index.js
@@ -83,6 +83,9 @@ var _throw = function(status, optDetail) {
   throw e;
 };
 
+Err.messages = ErrorMessages;
+Err._throw = _throw;
+
 var Bunzip = function(inputStream, outputStream, level) {
   this.writePos = this.writeCurrent = this.writeCount = 0;
 
@@ -105,15 +108,6 @@ Bunzip.prototype.validateFileStart = function() {
 
   return this._start_bunzip(level);
 };
-Bunzip.prototype._init_block = function() {
-  var moreBlocks = this._get_next_block();
-  if ( !moreBlocks ) {
-    this.writeCount = -1;
-    return false; /* no more blocks */
-  }
-  this.blockCRC = new CRC32();
-  return true;
-};
 /* XXX micro-bunzip uses (inputStream, inputBuffer, len) as arguments */
 Bunzip.prototype._start_bunzip = function(level) {
   /* Fourth byte (ascii '1'-'9'), indicates block size in units of 100k of
@@ -126,6 +120,15 @@ Bunzip.prototype._start_bunzip = function(level) {
   this.dbufSize = 100000 * level;
   this.nextoutput = 0;
   this.streamCRC = 0;
+};
+Bunzip.prototype._init_block = function() {
+  var moreBlocks = this._get_next_block();
+  if ( !moreBlocks ) {
+    this.writeCount = -1;
+    return false; /* no more blocks */
+  }
+  this.blockCRC = new CRC32();
+  return true;
 };
 Bunzip.prototype._get_next_block = function() {
   var i, j, k;
@@ -590,15 +593,17 @@ Bunzip.table = function(input, callback, multistream) {
           !inputStream.eof()) {
         // note that start_bunzip will also resync the bit reader to next byte
         bz._start_bunzip(inputStream, outputStream);
-        console.assert(bz.dbufSize === blockSize,
-                       "shouldn't change block size within multistream file");
+        console.assert(
+          bz.dbufSize === blockSize,
+          "shouldn't change block size within multistream file"
+        );
       } else break;
     }
   }
 };
 
 Bunzip.Stream = Stream;
-Bunzip.fdReadStream = require('./readstream');
+Bunzip.fdReadFile = require('./readfiledescriptor');
 Bunzip.fdWriteStream = require('./writestream');
 Bunzip.searchHexString = require('./searchhexstring');
 require('./readblock')(Bunzip);
